Add files via upload
This is the python implementation of Welford's with covariance calculation added. It shares the same four test vectors hardcoded into it as the SV testbench. This is also the update SV module for stats_accum that only computes the covariance matrix and uses the diagonal as the variance. This is a more sequential approach in order to save room on the FPGA. If needed a more parallel module can be added at a later date.
I am also uploading the testbench that I forgot to upload initially.